PSM: keep references to scene_ valid upon receiving full scenes (#2745)
plan_execution-related modules rely on `plan.planning_scene_` in many places
to point to the currently monitored scene (or a diff on top of it).

Before this patch, if the PSM would receive full scenes during execution,
`plan.planning_scene_` would not include later incremental updates anymore
because the monitor created a new diff scene.
---------

Co-authored-by: v4hn <me@v4hn.de>